% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAverage.R
\name{getAverage}
\alias{getAverage}
\title{Get Average Score For a Group}
\usage{
getAverage(column, choice, dataframe, subject)
}
\arguments{
\item{column}{a column of a dataframe (character), in "name of column" form. NOT a score list}

\item{dataframe}{a dataframe with at least a list of scores for some subject and students associated with the scores who can be divided into at least two groups (Ex: male/female)}

\item{subject}{any subject that has a list of scores attached, this example uses math/reading/writing/total, but it can be any subject as long as it's written properly and shows up in the dataframe. It must be in character form without the name of the dataframe and the $.}

\item{group}{a value in the aforementioned column (character)}
}
\description{
"getAverage" takes a dataframe, a column within that dataframe, a group within that column, and a subject with scores attached. It filters the dataframe for rows containing only that group, then finds the average of whatever subject was chosen and returns that. Note: To find the average total score, one must use a returned dataframe from getTotalScoreColumn() or have manually added up the three scores.
}
\examples{
getAverage(exams$gender, "male", exams, "math.score")
}
\keyword{SAT}
\keyword{average}
\keyword{calculate}
\keyword{collegeboard}
\keyword{exam}
\keyword{grading}
