% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTotalScoreColumn.R
\name{getTotalScoreColumn}
\alias{getTotalScoreColumn}
\title{Get Total Score Column}
\usage{
getTotalScoreColumn(
  mathScoreList,
  readingScoreList,
  writingScoreList,
  dataframe
)
}
\arguments{
\item{mathScoreList}{a list of math scores (numeric), in dataframe$scoreList form}

\item{readingScoreList}{a list of reading scores (numeric), in dataframe$scoreList form}

\item{writingScoreList}{a list of writing scores (numeric), in dataframe$scoreList form}

\item{dataframe}{a dataframe with at least three columns of numeric data (math/reading/writing) and a student associated with each}
}
\description{
"getTotalScoreColumn" takes a dataframe with three lists of scores (math, reading, writing) and adds them together, then returns a copy of the dataframe with "Total Score" column appended on. It first inputs the totals into a separate list, turns the values into numeric, and cbinds that list to the dataframe to avoid the values being input as type list.
}
\examples{
getTotalScoreColumn(exams$math.score, exams$reading.score, exams$writing.score, exams)
}
\keyword{SAT}
\keyword{calculate}
\keyword{collegeboard}
\keyword{exam}
\keyword{grading}
\keyword{total}
